lexer: int & float literals
Some slop ass code to tokenize int and float literals. Will clean up
later

+++ hello.c
@@ -2,9 +2,9 @@ extern int printf(const char* format, ...);
 
 int main() {
     int a;
-    int b;
+    float b;
     a = 65;
-    b = 66;
+    b = 66.0;
 
     if (a == b) {
         a + b;
